Add PREORDER exceptions in Spanish rules
PREORDER means that those exceptions are applied to the numbers
which immediately precede an order, such as millón, billón, etc.

The entry for 101 has been added to the lookup table so that it
will be treated as a simple number, rather than a compound one, when
it comes time to apply PREORDER exceptions.