Extract sections from Readme into separate documents
This is done in order to separate user documentation from developer
documentation. Everything user needs to know can be found in Readme or
by running `help(numspell)` in the Python interactive prompt.

The contents of the 'doc' directory is targeted at developers willing to
port the library, extend it or contribute in some other way to it.